Made IndigoAtomic actually use an atomic add. Made ThreadSafeRefCounted use IndigoAtomic.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@6110 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/MyThread.cpp
@@ -59,7 +59,7 @@ threadFunction(void* the_thread_)
 	the_thread->run();
 
 	// Decrement the reference count
-	int new_ref_count = the_thread->decRefCount();
+	int64 new_ref_count = the_thread->decRefCount();
 	assert(new_ref_count >= 0);
 	if(new_ref_count == 0)
 		delete the_thread;
